pyproject.toml
[project]
name = "base-models-reasoning-interp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "accelerate>=1.10.1",
    "aiohappyeyeballs>=2.6.1",
    "aiohttp>=3.12.15",
    "aiosignal>=1.4.0",
    "annotated-types>=0.7.0",
    "anthropic>=0.64.0",
    "async-timeout>=5.0.1",
    "attrs>=25.3.0",
    "automated-interpretability>=0.0.15",
    "babe>=0.0.7",
    "beartype<0.15.0",
    "better-abc>=0.0.3",
    "bidict>=0.23.1",
    "blobfile>=2.1.1,<3.0.0",
    "boostedblob>=0.15.6",
    "certifi>=2025.8.3",
    "charset-normalizer>=3.4.3",
    "chat-limiter",
    "click>=8.2.1",
    "cloudpickle>=3.1.1",
    "config2py>=0.1.38",
    "contourpy>=1.3.3",
    "datasets>=2.2.1",
    "diffusers>=0.35.1",
    "dill<0.3.9",
    "docker-pycreds>=0.4.0",
    "docstring-parser>=0.17.0",
    "dol>=0.3.19",
    "einops>=0.8.1",
    "fancy-einsum>=0.0.3",
    "farama-notifications>=0.0.4",
    "filelock>=3.19.1",
    "fonttools>=4.59.2",
    "frozenlist>=1.7.0",
    "fsspec<=2025.3.0",
    "gdown>=5.2.0",
    "gitdb>=4.0.12",
    "gitpython>=3.1.45",
    "gprof2dot>=2025.4.14",
    "graze>=0.1.29",
    "h11>=0.16.0",
    "h5py>=3.14.0",
    "hf-transfer>=0.1.9",
    "httpcore>=1.0.9",
    "httpx<0.28.0",
    "i2>=0.1.48",
    "idna>=3.10",
    "imageio>=2.37.0",
    "imageio-ffmpeg>=0.6.0",
    "importlib-resources>=6.5.2",
    "iniconfig>=2.1.0",
    "jaxtyping>=0.3.2",
    "jinja2>=3.1.6",
    "jiter>=0.10.0",
    "joblib>=1.5.2",
    "kiwisolver>=1.4.9",
    "markdown-it-py>=4.0.0",
    "markupsafe>=3.0.2",
    "matplotlib>=3.10.6",
    "mdmm>=0.1.3",
    "mdurl>=0.1.2",
    "mpmath>=1.3.0",
    "msgspec>=0.19.0",
    "multidict>=6.6.4",
    "multiprocess<0.70.17",
    "networkx>=3.5",
    "nltk>=3.9.1",
    "nnsight>=0.5.1",
    "numpy>=1.26.0,<2.0.0",
    "openai>=1.102.0",
    "orjson>=3.11.3",
    "pillow>=11.3.0",
    "plotly>=5.24.1",
    "plotly-express>=0.4.1",
    "pluggy>=1.6.0",
    "propcache>=0.3.2",
    "protobuf>=6.32.0",
    "psutil>=7.0.0",
    "pyarrow>=21.0.0",
    "pycryptodomex>=3.23.0",
    "pydantic>=2.11.7",
    "pydantic-core>=2.33.2",
    "pyparsing>=3.2.3",
    "pytest>=8.4.1",
    "pytest-profiling>=1.8.1",
    "python-dotenv>=1.1.1",
    "python-engineio>=4.12.2",
    "python-socketio>=5.13.0",
    "pytz>=2025.2",
    "pyzmq>=27.0.2",
    "regex>=2025.8.29",
    "requests>=2.32.5",
    "rich>=14.1.0",
    "rtree>=1.4.1",
    "sae-lens>=6.10.0",
    "scikit-learn>=1.7.1",
    "seaborn>=0.13.2",
    "sentencepiece>=0.2.1",
    "sentry-sdk>=2.35.2",
    "setproctitle>=1.3.6",
    "shellingham>=1.5.4",
    "simple-parsing>=0.1.7",
    "simple-websocket>=1.1.0",
    "smmap>=5.0.2",
    "sniffio>=1.3.1",
    "soupsieve>=2.8",
    "sympy>=1.14.0",
    "tabulate>=0.9.0",
    "tenacity>=9.1.2",
    "threadpoolctl>=3.6.0",
    "tokenizers>=0.22.0",
    "toml>=0.10.2",
    "tomli>=2.2.1",
    "transformer-lens>=2.16.1",
    "transformers>=4.56.0",
    "transforms3d>=0.4.2",
    "trimesh>=4.7.4",
    "typeguard>=4.4.4",
    "typer<0.13.0",
    "tzdata>=2025.2",
    "urllib3>=2.5.0",
    "uvloop>=0.21.0",
    "vllm>=0.1.2",
    "wandb>=0.21.3",
    "websocket-client>=1.8.0",
    "wsproto>=1.2.0",
    "xxhash>=3.5.0",
    "yarl>=1.20.1",
    "zstandard>=0.24.0",
]

[tool.uv.sources]
chat-limiter = { git = "https://github.com/FlyingPumba/chat-limiter.git" }
